requirements.txt
absl-py==1.4.0
adbutils==2.7.1
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
airtest==1.2.7
allure-pytest==2.13.2
allure-python-commons==2.13.2
alphalens==0.4.0
altair==4.2.2
altgraph==0.17.3
apkutils2==1.0.0
appdirs==1.4.4
APScheduler==3.9.1.post1
aria2rpc==1.1.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
av==8.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
blinker==1.5
bokeh==3.1.0
bottle==0.12.25
briefcase==0.3.16
bs4==0.0.1
build==1.0.3
cached-property==1.5.2
cachetools==5.3.0
cattrs==22.2.0
cchardet==2.1.7
cefpython3==66.1
certifi==2022.12.7
cffi==1.15.1
cfscrape==2.1.1
chardet==5.2.0
charset-normalizer==3.0.1
cheerio==1.4
chinese-calendar==1.8.0
cigam==0.0.3
click==8.1.3
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.1.14
configobj==5.0.8
contourpy==1.0.7
control==0.9.4
cookiecutter==2.4.0
cors==1.0.1
crcmod==1.7
cssselect==1.2.0
cycler==0.11.0
ddddocr==1.5.5
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
distlib==0.3.6
Django==4.2.13
empyrical==0.5.5
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
facebook-wda==1.4.6
fake-useragent==1.1.1
filelock==3.10.0
filestools==0.1.3
filetype==1.2.0
Flask==3.0.3
flaskwebgui==1.1.1
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
greenlet==2.0.1
grequests==0.7.0
grpcio==1.54.0
gTTS==2.3.1
h11==0.14.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hrpc==1.0.9
hstspreload==2024.5.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
icon==0.0.5
idna==3.4
image==1.5.33
img==2.5
importlib==1.0.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.10.0
itsdangerous==2.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
Js2Py==0.74
jsonpath==0.82
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
loguru==0.6.0
lxml @ file:///C:/Users/rjcsyb2/Desktop/workcard-master/lxml-4.9.3-cp38-cp38-win_amd64.whl#sha256=92af161ecbdb2883c4593d5ed4815ea71b31fafd7fd05789b23100d081ecac96
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==6.1.0
multidict==6.0.4
Naked==0.1.32
networkx==3.0
nose==1.3.7
Nuitka==2.3.11
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.18.0
openai==0.27.2
opencv-contrib-python==4.7.0.68
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.84
openpyxl==3.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.2.0
packaging==23.0
pandas==1.5.2
pandas-datareader==0.10.0
parso==0.8.3
patsy==0.5.3
pefile==2022.5.30
pickleshare==0.7.5
pillow==10.3.0
pipenv==2018.10.13
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
playwright==1.33.0
pluggy==1.0.0
prettytable==3.6.0
prompt-toolkit==3.0.37
protobuf==3.20.3
proxy_tools==0.1.0
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
py==1.11.0
py-mini-racer==0.6.0
py2exe==0.13.0.1
pyarmor==8.5.10
pyarmor.cli.core==6.5.3
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.0
pyecharts==2.0.1
pyee==9.0.4
pyelftools==0.29
PyExecJS==1.5.1
pygame==2.1.3.dev8
PyGameAuto==1.6.8
PyGameAutoAndroid==0.0.22
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyjsparser==2.7.1
pyminitouch==0.3.3
Pympler==1.0.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypinyin==0.48.0
pypiwin32==223
pyproject_hooks==1.0.0
pyquery==2.0.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
PySide2==5.15.2.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.1
python-dateutil==2.8.2
python-lzo==1.15
python-slugify==6.1.2
pythonnet==3.0.3
pytweening==1.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywebview==5.1
pywin32==305
pywin32-ctypes==0.2.0
pywinauto==0.6.3
PyYAML==6.0
readmdict==0.1.1
reload==0.9
requests==2.28.2
requests-cache==0.9.8
requests-file==2.1.0
requests-oauthlib==1.3.1
retry==0.9.2
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.0
semver==2.13.0
shellescape==3.8.1
shiboken2==5.15.2.1
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spark-parser==1.8.9
sqlparse==0.5.0
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.20.0
streamlit-bokeh-events==0.1.2
subprocess32==3.5.4
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tesseract==0.1.3
text-unidecode==1.3
threadpoolctl==3.1.0
tinyaes==1.0.4
tkintertools==2.6.21.1
tkwebview2==3.5.0
tldextract==5.1.2
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
ttkbootstrap==1.10.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uiautomator==1.0.2
uiautomator2==3.0.18
uncompyle6==3.9.0
url-normalize==1.4.3
urllib3==1.26.14
validators==0.20.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchdog==2.3.1
wcwidth==0.2.6
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==2.3.4
whichcraft==0.6.1
win32-setctime==1.1.0
WMI==1.5.1
wrapt==1.14.1
wsproto==1.2.0
xdis==6.0.5
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.13.0
xyzservices==2023.2.0
yarl==1.8.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
zstandard==0.22.0
